+++ plug-n-code/src/app/components/toolbar/toolbar.component.html
@@ -1,5 +1,5 @@
 <div id="container">
-  <button *ngIf="isMobile" class="basic blue small ui button icon" (click)="isNavOpenChange.emit(true)">
+  <button *ngIf="isMobile" id="open-nav" class="basic blue small ui button icon" (click)="isNavOpenChange.emit(true)">
     <i class="sidebar icon"></i>
   </button>
